Update Day 10 Project Calculator.py

+++ 100 Projects/Day 10 Project/Day 10 Project Calculator.py
@@ -22,13 +22,13 @@ operations = {
 
 def calculator():
     print(logo)
-    num1 = int(input("Enter the first number: "))
+    num1 = float(input("Enter the first number: "))
     for operation in operations:
         print(operation)
     cont = True
     while (cont):
         operation_symbol = input("Pick an operation: ")
-        num2 = int(input("Enter the next number: "))
+        num2 = float(input("Enter the next number: "))
         operation_function = operations[operation_symbol]
         result = operation_function(num1, num2)
         print(f"{num1} {operation_symbol} {num2} = {result}")
